realtime_object_detector/requirements.txt
autopep8==1.5
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cycler==0.10.0
Cython==0.29.14
gTTS==2.1.1
gTTS-token==1.1.3
idna==2.8
kiwisolver==1.1.0
matplotlib==3.1.2
numpy==1.18.1
pycodestyle==2.5.0
pydub==0.23.1
pyparsing==2.4.6
python-dateutil==2.8.1
python-openal==0.4.0
requests==2.22.0
scipy==1.4.1
six==1.14.0
soupsieve==1.9.5
urllib3==1.25.8
